# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoRecicla Bogotá API - Recycling Management System for Residential Complexes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all recyclable materials
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialListResponse"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a recyclable material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # COLLECTION EVENTS
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all collection events
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: complexId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a collection event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # RECYCLING RECORDS
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
  /profile/photo:
    post:
      operationId: uploadProfilePhoto
      tags: [profile]
      summary: Upload profile photo URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "200":
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, documentNumber, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        documentNumber:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: string
        apartment:
          type: ["string", "null"]
        complexId:
          type: ["integer", "null"]
        complexName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, resident]
        status:
          type: string
          enum: [active, inactive]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Material:
      type: object
      required: [id, name, description, recyclingInstructions, binColor, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        recyclingInstructions:
          type: string
        binColor:
          type: string
        createdAt:
          type: string
    MaterialInput:
      type: object
      required: [name, description, recyclingInstructions, binColor]
      properties:
        name:
          type: string
        description:
          type: string
        recyclingInstructions:
          type: string
        binColor:
          type: string
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        recyclingInstructions:
          type: string
        binColor:
          type: string
    MaterialListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Material"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # EVENT
    Event:
      type: object
      required: [id, eventName, complexId, date, hour, responsiblePerson, location, status, createdAt]
      properties:
        id:
          type: integer
        eventName:
          type: string
        complexId:
          type: integer
        complexName:
          type: ["string", "null"]
        date:
          type: string
        hour:
          type: string
        responsiblePerson:
          type: string
        location:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [eventName, complexId, date, hour, responsiblePerson, location, status]
      properties:
        eventName:
          type: string
        complexId:
          type: integer
        date:
          type: string
        hour:
          type: string
        responsiblePerson:
          type: string
        location:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
    EventUpdate:
      type: object
      properties:
        eventName:
          type: string
        complexId:
          type: integer
        date:
          type: string
        hour:
          type: string
        responsiblePerson:
          type: string
        location:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
    EventListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # RECYCLING RECORD
...
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        apartment:
          type: string
        email:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    PhotoInput:
      type: object
      required: [photoUrl]
      properties:
        photoUrl:
          type: string